Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -1,20 +1,29 @@
 package com.educandoweb.course.services;
 
+import com.educandoweb.course.config.TestConfig;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DataBaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
 
+	private final TestConfig testConfig;
 	@Autowired
 	private UserRepository userRepository;
+
+	UserService(TestConfig testConfig) {
+		this.testConfig = testConfig;
+	}
 
 	public List<User> findAll() {
 		return userRepository.findAll();
@@ -30,7 +39,16 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if(userRepository.existsById(id)){
+				userRepository.deleteById(id);
+			}
+			else {
+				throw new ResourceNotFoundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
